requirements.txt
PyGithub
jsonschema
toml